Simplify ptrace call by removing unnecessary conversions
The original code converted the `value` to a byte array and then
marshaled this array to an unmanaged memory address before calling
the `ptrace` function. The updated code removes these steps and
directly passes the `value` to the `ptrace` function, simplifying
the process of writing the value to the target memory address.

Refactor long value memory write process

Simplified the process of writing a long value to a target memory address by directly casting the long value to an `nint` type. This eliminates the need for converting the long value to a byte array and using `Marshal.UnsafeAddrOfPinnedArrayElement`. Updated the `ptrace` function call to use the `targetPointer` and the directly cast `data` value.